nice

+++ philo/src/routine_utils.c
@@ -6,24 +6,24 @@
 /*   By: tsofien- <tsofien-@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/03 10:58:49 by tsofien-          #+#    #+#             */
-/*   Updated: 2024/09/16 22:38:58 by tsofien-         ###   ########.fr       */
+/*   Updated: 2024/09/17 11:49:25 by tsofien-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-void	philo_msg(MSG msg, long time, int position)
+void	philo_msg(MSG msg, long long time, int position)
 {
 	if (msg == DEAD)
-		printf("%ld ms %d died\n", time, position);
+		printf("%lld ms %d died\n", time, position);
 	else if (msg == FORK)
-		printf("%ld ms %d has taken a fork\n", time, position);
+		printf("%lld ms %d has taken a fork\n", time, position);
 	else if (msg == EAT)
-		printf("%ld ms %d is eating\n", time, position);
+		printf("%lld ms %d is eating\n", time, position);
 	else if (msg == SLEEP)
-		printf("%ld ms %d is sleeping\n", time, position);
+		printf("%lld ms %d is sleeping\n", time, position);
 	else if (msg == THINK)
-		printf("%ld ms %d is thinking\n", time, position);
+		printf("%lld ms %d is thinking\n", time, position);
 }
 
 long long	get_current_time(void)
